+++ lualib/c-stdlib/mocked_math.h
