+++ lualib/c-stdlib/mocked_stdio.h
@@ -5,7 +5,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-#include "../include/ckb_cell_fs.h"
+#include "../../include/ckb_cell_fs.h"
 
 #define BUFSIZ 512
 #define EOF (-1)
@@ -53,14 +53,10 @@ int sprintf(char *__s, const char *__format, ...);
 
 int vfprintf(FILE *__s, const char *__format, ...);
 
-int vprintf(const char *__format, ...);
-
 int vsprintf(char *__s, const char *__format, ...);
 
 int snprintf(char *__s, size_t __maxlen, const char *__format, ...);
 int snprintf_(char *__s, size_t __maxlen, const char *__format, ...);
-
-int vsnprintf(char *__s, size_t __maxlen, const char *__format, ...);
 
 int fscanf(FILE *__stream, const char *__format, ...);
