+++ lualib/c-stdlib/src/math_log.c
