+++ lualib/c-stdlib/src/math_pow.c
