+++ lualib/c-stdlib/src/math_private.h
